# Less locking for metric vectors.