@beorn7 2015-11-09
Thanks for your contribution.

I like the general idea. I'll have a closer look ASAP.