@beorn7 2015-11-10
:+1: Looks perfect to me. Merging...

And thanks again. I guess we can apply the same idea to a few other performance critical paths...